And Move out split container as well
